requirements.txt
amqp==5.3.1
asgiref==3.8.1
axios==0.4.0
billiard==4.2.1
boto3==1.40.71
botocore==1.40.71
celery==5.5.3
certifi==2025.1.31
cffi==1.17.1
channels==4.2.2
charset-normalizer==3.4.1
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap4==2024.10
cryptography==44.0.2
dj-rest-auth==7.0.1
Django==5.1.7
django-allauth==65.11.1
django-anymail==13.1
django-appconf==1.1.0
django-axes==8.0.0
django-celery==3.1.17
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-cryptography==1.1
django-filter==25.1
django-jsonfield==1.4.1
django-oauth-toolkit==3.0.1
django-oauth2-provider==0.2.6.1
django-redis==6.0.0
django-storages==1.14.6
django-stripe-payments==2.0.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
geographiclib==2.0
geopy==2.4.1
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.6
graphql-relay==3.2.0
gunicorn==20.1.0
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
jmespath==1.0.1
jwcrypto==1.5.6
kombu==5.5.4
lxml==5.3.2
markdown-it-py==3.0.0
mdurl==0.1.2
mysql-connector==2.2.9
mysqlclient==2.2.7
oauthlib==3.3.1
packaging==25.0
parameterized==0.9.0
pillow==11.2.1
pluggy==1.6.0
promise==2.3
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pyactiveresource==2.2.2
pycparser==2.22
Pygments==2.19.1
PyJWT==2.9.0
PyMySQL==1.1.2
pytest==8.3.5
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
reportlab==4.4.3
requests==2.32.3
rich==14.0.0
s3transfer==0.14.0
setuptools==80.9.0
ShopifyAPI==12.7.0
shortuuid==1.0.13
six==1.17.0
sqlparse==0.5.3
stripe==12.0.0
text-unidecode==1.3
typing_extensions==4.13.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
